First, we add a .gitignore file set up for Python development.  An excellent base version of this file can be found on GitHub: https://github.com/github/gitignore/blob/master/Python.gitignore
To the base .gitignore file we add a line to ignore sqlite database files. These files should be considered transitory artifacts of the development process and as such should not be committed to version control.

Next, we will update the Python software dependencies that our app will require going forward. By listing dependencies in the `setup.py` script at the root of our application package, we can ensure that installing our package will also result in these additional packages being installed.

Finally, we set our application configuration to include the configuration for the jinja2 add-on. This replaces the default "chameleon" templating add-on that comes out-of-the-box with the `sqlalchemy` scaffold.