First, we prepare stub view methods for the endpoints we'll need to serve.
At the same time, we create stub templates to stand in for the templates we'll eventually need to flesh out.  This allows us to verify that our routes are properly working before we attempt to go any further

We should also remove the unused template created by our `sqlalchemy` scaffold.  This keeps our project directory clean of unused code.