We now add a few methods on the entry model that will support our view code.
It's good practice to attach code that directly references or manipulates a model to that model class rather than implementing it inline in view code.

In addition, we create a class that will support pagination for our blog entries.  The purpose of the class is to create URLs for pages of blog entries.