In order to be able to add or edit posts, we must add forms.
These will facilitate the communication of data from the client back to the server.

Next, we add views for creating and editing blog posts.  Use the forms created earlier to manage data transferred by the request from the client to the server.

In addition, we update the template for the new create/edit views so it will render our form