We need to add security for creating and editing blog entries. We begin by defining a security factory for the Entry class that will provide an ACL for entries
Next, we update application configuration to use the new ACL we have created.

Finally, we require the 'create' or 'edit' permission for the corresponding views using a new argument to the `@view_config` decorator.

Our security ACL for entries ensures that the required permission is only available to a user who is authenticated, securing our blog from public access.