We start our view work by creating the code to view a paginated list of blog entries and the code for viewing an individual blog entry.
We also need to add a template that will provide the entire site-wide layout.  Our other templates will extend this one.

Next, We create the template that will show us a single blog entry.

And finally, We create the template that will show us a paginated list of blog entries.