Now that creating and editing entries is secured, we need to be able to log in to our application to make any further entries.
Adding a login form will enable simple rendering of the login page when required.

The User model class should be responsible for verifying a user's password.  And we want a controller method that will fetch a user given a name.

With all that in place, We create the view for logging in to the application.

In addition, we modify the index page view so that if we are *not* logged in, we will have a login form to render.

We also update the edit view to provide information to the template about whether or not a user is logged in.  This will allow showing/hiding the "edit" button in the template.

Finally, We update the blog index page template so that it will show the login form if present.  If the form is not present, we must be logged in, and we can show the "create" button.

The single blog entry view template is similarly updated to show the *edit* button only if we are logged in.