We can create a simple template filter for jinja2 templates that will render text that is in Markdown format as HTML.  This will allow for prettier formatting for our blog entries.  By enabling the `codehilite` and `fenced_block` extensions to the markdown filter, we can have syntax hilighting similar to that used in github.
We must update our development and production configuration files to include this new filter.  We could do it in the `__init__.py` file but here is also fine.

Finally, we have to update our display template to use the new filter we created.

Adding css rules for colorization (generated by Pygments in the style of github colorization) will give us a nice-looking result for code blocks.