Add a services package
We will put code here that will provide a unified interface for working with our data model.

Creating a folder with an __init__.py file in it creates a Python 'package', which can hold other Python files.

Using Packages allows us to organize our code in a way that makes it easy to find what we need to work with.