add an index view
This view will display a list of blog records.

The list will be paginated.

if there is no list, we'll see a statement to that effect.

There will also be a link to the page for creating a new entry

At this point, we should be able to start our application and see the home page. Try loading http://localhost:6543