Update the edit and create views for blog records
The create view uses our BlogCreateForm to validate data from an incoming request and transfer it to a new BlogRecord object.

The edit view uses the blog service to retrieve an existing entry by ID, and then updates that with data from a request validated by the BlogUpdateForm

Both form views only process data that has been passed in a POST request.