7 -- Add authorization for views
An authorization system allows us to require a 'permission' in order to reach certain views.

We add a permission declaration for those views we wish to protect. Then users who do not have that permission will be refused access.  Those who do will be allowed.

We must configure Authorization at the same time, but we will implement that in our next step.