Add login form to index page
We can use the  attribute of the Pyramid request object to determine if a user is authenticated.

When the value is true (there is an id) then a user has logged in.  We can show them a welcome message and a log out button.

When the value is false (no id is present) then the current user has not logged in.  So we show them a login form.