8 -- Add login/logout view
This view will look up a user object using the 'username' passed in from the login form on a POST request.
It will then verify the password using the API we've added to the User model.
If a user object is found, and the password matches, then the 'remember' method is used to write the HTTP Headers needed to persist the fact that we have an authenticated user.
The 'authorization policy' we registered earlier is responsible for doing this work, so how exactly login is persisted depends on that policy.

If no user is found, if the password does not match, or if we have reached this view with the 'out' action (a GET request), then we will use the 'forget' API to remove authentication information.
Again, our registered authentication policy is responsible for this action.