10 -- Hide edit link from others
By adding the BlogRecordFactory as the factory for the blog view as well, we can then check for the edit permission in the context of a template.
If the current user is the author of the post we are viewing, then they will be able to see the 'edit' button.
Otherwise, it will be hidden from them